Move contact page constants, types and utils to separate files
