+++ src/domain/events/eventList/EventList.tsx
@@ -131,7 +131,7 @@ const EventList: React.FC<EventListProps> = ({
 
 const EventListContainer: React.FC<EventListContainerProps> = (props) => {
   const { baseVariables, className, listType, setListType, skip } = props;
-  const eventListId = uniqueId('event-list-');
+  const eventListId = React.useRef<string>(uniqueId('event-list-')).current;
   const { t } = useTranslation();
   const history = useHistory();
   const location = useLocation<EventsLocationState>();
