+++ src/domain/help/pages/ContactPage.tsx
@@ -1,19 +1,9 @@
-import {
-  Field,
-  Form,
-  Formik,
-  FormikErrors,
-  FormikHelpers,
-  FormikTouched,
-} from 'formik';
-import forEach from 'lodash/forEach';
-import set from 'lodash/set';
+import { Field, Form, Formik, FormikHelpers } from 'formik';
 import uniqueId from 'lodash/uniqueId';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 import { scroller } from 'react-scroll';
-import * as Yup from 'yup';
 
 import Button from '../../../common/components/button/Button';
 import TextAreaField from '../../../common/components/formFields/TextAreaField';
@@ -24,102 +14,20 @@ import {
   usePostFeedbackMutation,
   usePostGuestFeedbackMutation,
 } from '../../../generated/graphql';
-import { VALIDATION_MESSAGE_KEYS } from '../../app/i18n/constants';
 import PageWrapper from '../../app/layout/PageWrapper';
 import { authenticatedSelector } from '../../auth/selectors';
+import {
+  CONTACT_FORM_FIELD,
+  initialValues,
+  validationSchema,
+} from '../constants';
+import { ContactFormFields } from '../types';
+import { scrollToFirstError, showErrors } from '../utils';
 import styles from './contactPage.module.scss';
 
-enum CONTACT_FORM_FIELD {
-  BODY = 'body',
-  EMAIL = 'email',
-  NAME = 'name',
-  SUBJECT = 'subject',
-}
-
-type ContactFormFields = {
-  [CONTACT_FORM_FIELD.BODY]: string;
-  [CONTACT_FORM_FIELD.EMAIL]: string;
-  [CONTACT_FORM_FIELD.NAME]: string;
-  [CONTACT_FORM_FIELD.SUBJECT]: string;
-};
-
-const initialValues = {
-  [CONTACT_FORM_FIELD.BODY]: '',
-  [CONTACT_FORM_FIELD.EMAIL]: '',
-  [CONTACT_FORM_FIELD.NAME]: '',
-  [CONTACT_FORM_FIELD.SUBJECT]: '',
-};
-
-export const validationSchema = Yup.object().shape({
-  [CONTACT_FORM_FIELD.NAME]: Yup.string().required(
-    VALIDATION_MESSAGE_KEYS.STRING_REQUIRED
-  ),
-  [CONTACT_FORM_FIELD.EMAIL]: Yup.string()
-    .required(VALIDATION_MESSAGE_KEYS.STRING_REQUIRED)
-    .email(VALIDATION_MESSAGE_KEYS.EMAIL),
-  [CONTACT_FORM_FIELD.SUBJECT]: Yup.string().required(
-    VALIDATION_MESSAGE_KEYS.STRING_REQUIRED
-  ),
-  [CONTACT_FORM_FIELD.BODY]: Yup.string().required(
-    VALIDATION_MESSAGE_KEYS.STRING_REQUIRED
-  ),
-});
-
-export const scrollToFirstError = ({
-  error,
-}: {
-  error: Yup.ValidationError;
-}): void => {
-  forEach(error.inner, (e) => {
-    const fieldId = e.path;
-    const field = document.getElementById(fieldId);
-
-    /* istanbul ignore else */
-    if (field) {
-      scroller.scrollTo(fieldId, {
-        delay: 0,
-        duration: 500,
-        offset: -200,
-        smooth: true,
-      });
-
-      field.focus();
-      return false;
-    }
-  });
-};
-
-export const showErrors = ({
-  error,
-  setErrors,
-  setTouched,
-}: {
-  error: Yup.ValidationError;
-  setErrors: (errors: FormikErrors<ContactFormFields>) => void;
-  setTouched: (
-    touched: FormikTouched<ContactFormFields>,
-    shouldValidate?: boolean
-  ) => void;
-}): void => {
-  /* istanbul ignore else */
-  if (error.name === 'ValidationError') {
-    const newErrors = error.inner.reduce(
-      (acc, e: Yup.ValidationError) => set(acc, e.path, e.errors[0]),
-      {}
-    );
-    const touchedFields = error.inner.reduce(
-      (acc, e: Yup.ValidationError) => set(acc, e.path, true),
-      {}
-    );
-
-    setErrors(newErrors);
-    setTouched(touchedFields);
-    scrollToFirstError({ error });
-  }
-};
-
 const ContactPage: React.FC = () => {
-  const successId = uniqueId('contact-form-success-');
+  const successId = React.useRef<string>(uniqueId('contact-form-success-'))
+    .current;
   const { t } = useTranslation();
   const authenticated = useSelector(authenticatedSelector);
 
@@ -144,9 +52,17 @@ const ContactPage: React.FC = () => {
         await postGuestFeedback({ variables: { input: values } });
       }
 
-      setSuccess(true);
       resetForm();
-      validateForm();
+      await validateForm();
+
+      scroller.scrollTo(successId, {
+        duration: 300,
+        offset: -200,
+        smooth: true,
+      });
+      setSuccess(true);
+
+      document.getElementById(CONTACT_FORM_FIELD.NAME)?.focus();
     } catch (e) {
       setSuccess(false);
     }
@@ -194,10 +110,24 @@ const ContactPage: React.FC = () => {
                 setErrors,
                 setTouched,
               });
+
+              scrollToFirstError({ error });
             }
           };
+
           return (
             <Form className={styles.contactForm} noValidate>
+              <div id={successId}>
+                {success && (
+                  <Notification
+                    label={t('helpPage.contactPage.titleSuccess')}
+                    type="success"
+                  >
+                    {t('helpPage.contactPage.textSuccess')}
+                  </Notification>
+                )}
+              </div>
+
               <h2>{t('helpPage.contactPage.titleContactInfo')}</h2>
               <FormGroup>
                 <Field
@@ -237,22 +167,9 @@ const ContactPage: React.FC = () => {
                 />
               </FormGroup>
 
-              <FormGroup>
-                <Button onClick={() => handleSubmit()} fullWidth>
-                  {t('helpPage.contactPage.buttonSend')}
-                </Button>
-              </FormGroup>
-
-              <div id={successId}>
-                {success && (
-                  <Notification
-                    label={t('helpPage.contactPage.titleSuccess')}
-                    type="success"
-                  >
-                    {t('helpPage.contactPage.textSuccess')}
-                  </Notification>
-                )}
-              </div>
+              <Button onClick={() => handleSubmit()} fullWidth>
+                {t('helpPage.contactPage.buttonSend')}
+              </Button>
             </Form>
           );
         }}
